pyproject.toml
[tool.poetry]
name = "hackerspace-notify-v2"
version = "0.1.0"
description = ""
authors = ["bernes1 <37117219+bernes1@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
paho-mqtt = "^2.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
